Trimmed Down llvm-toolchain's Dependencies
We do not require using prebuilt ninja, cmake.
Plus we are not building llvm on x86, therefore we do not require the
darwin and x86 counterparts of libraries we are going to use.

We are still picking libraries from a prebuilt GCC, but this is going
to change once we get over with the mess of compiling Android's GCC
manually on aarch64.

Signed-off-by: Gunwant Jain <therealgunwant@gmail.com>